requirements.txt
requests
urllib3<2
python-telegram-bot[job-queue]==20.8
schedule
flask
waitress
python-dotenv
telegram
deep-translator
pytrends
pandas
lxml
openai>=1.0.0
